rename controller for creating sites and create new one for site activities
