+++ src/device-registry/routes/api.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
 const deviceController = require("../controllers/create-device");
-const siteController = require("../controllers/manage-site");
+const siteController = require("../controllers/create-site");
 const middlewareConfig = require("../config/router.middleware");
 const componentController = require("../controllers/create-component");
 const eventController = require("../controllers/create-event");
