+++ src/device-registry/utils/create-activity.js
